+++ Panels/WizardPanel.js
@@ -100,6 +100,20 @@ const WizardPanelBase = kind({
 		onChange: PropTypes.func,
 
 		/**
+		 * Called when a transition completes.
+		 *
+		 * @type {Function}
+		 */
+		onTransition: PropTypes.func,
+
+		/**
+		 * Called before a transition begins.
+		 *
+		 * @type {Function}
+		 */
+		onWillTransition: PropTypes.func,
+
+		/**
 		* The text for previous button.
 		*
 		* @type {String}
@@ -163,10 +177,20 @@ const WizardPanelBase = kind({
 
 				onChange({index: prevIndex});
 			}
+		},
+		onTransition: (ev, {index, onTransition}) => {
+			if (onTransition) {
+				onTransition({index});
+			}
+		},
+		onWillTransition: (ev, {index, onWillTransition}) => {
+			if (onWillTransition) {
+				onWillTransition({index});
+			}
 		}
 	},
 
-	render: ({buttons, children, footer, index, total, nextButtonText, noAnimation, onIncrementStep, onDecrementStep, prevButtonText, reverseTransition, subtitle, title, ...rest}) => {
+	render: ({buttons, children, footer, index, total, nextButtonText, noAnimation, onIncrementStep, onDecrementStep, onTransition, onWillTransition, prevButtonText, reverseTransition, subtitle, title, ...rest}) => {
 		return (
 			<Panel {...rest}>
 				<Header
@@ -207,6 +231,8 @@ const WizardPanelBase = kind({
 							<ViewManager
 								arranger={SlideLeftArranger}
 								duration={400}
+								onTransition={onTransition}
+								onWillTransition={onWillTransition}
 								noAnimation={noAnimation}
 								reverseTransition={reverseTransition}
 							>